requirements.txt
Flask
Flask-Admin
Flask-SQLAlchemy
Flask-Stormpath
Flask-WTF
blinker
psycopg2
gunicorn